ref: replace API for more explicit one
Before I had decided to have `forest cmd`. The problem with this
approach is that it quickly got confusing and limiting. What is "clone"
for, "why does it act on roots and create acts on trees"?

To solve that, I've decided to have 2 main commands and proceed with
similar subcommands:

forest (roots|trees) <cmd>